Karakter Eklendi ve bazı işlemler yapıldı
karakter ekledim ama script eklenmiyor future/wck-014'ü incelerseniz harika olur
