--- Assets/CameraController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6cd3f1b056991460f83fb8982e510844
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
