--- Assets/CharacterScript.cs
@@ -1,16 +1,0 @@
-using UnityEngine;
-
-public class CharacterScript : MonoBehaviour
-{
-    public float speed = 5f;
-    public Camera camera;
-    void Update()
-    {
-        float horizontal = Input.GetAxis("Horizontal");
-        float vertical = Input.GetAxis("Vertical");
-
-        Vector3 movement = new Vector3(horizontal, 0f, vertical) * speed * Time.deltaTime;
-        transform.Translate(movement);
-        
-    }
-}
