--- Assets/CharacterScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cc6030a4e72334b87be222020376cb96
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
