+++ Assets/genSuit.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9c688e438394f4cfe98ecece7eee855b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
