+++ Assets/genSuit/Materials.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 154132e3dc50049a5872c4af1bcc593c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
