+++ Assets/genSuit/Materials/genSuit.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4ef944738df6cd34285471b45a6884a5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
