+++ Assets/genSuit/Textures.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7f7aab8b5581e44a4b740e7ba2958661
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
