+++ Assets/genSuit/Textures/Beige.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ef3be064f67f44104810b5353fbcd841
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
