+++ Assets/genSuit/Textures/Black.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 072946c0156f9407aac77308d2bfb9c8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
