+++ Assets/genSuit/Textures/Blue.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 478ef2e6190c644158b0b909b931299f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
