+++ Assets/genSuit/Textures/Dark Blue.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 772bd3088c3664fe28b1edf7310240d8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
