+++ Assets/genSuit/Textures/Striped.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 71bd157452cbe4482ade41cdae3e3769
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
