+++ Assets/genSuit/genSuit.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b7d9a5b3fad134b4689e06aafa3028d7
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
