+++ Assets/genSuit/genSuitSettings.lighting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f9a2258bea54f4563ab904551266badb
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 4890085278179872738
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
